Refactor to lib + small app.
